update basket and gift

+++ Controllers/UserController.cs
@@ -3,7 +3,9 @@ using ChineseRaffleApi.Dto;
 using ChineseRaffleApi.Models;
 using ChineseRaffleApi.Repository.DI;
 using ChineseRaffleApi.Services.DI;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace ChineseRaffleApi.Controllers
 {
@@ -18,6 +20,7 @@ namespace ChineseRaffleApi.Controllers
             _userService = userService;
         }
 
+        [Authorize(Roles = "Admin")]
         [HttpGet]
         public async Task<IActionResult> GetAllUsers()
         {
@@ -25,6 +28,7 @@ namespace ChineseRaffleApi.Controllers
             return Ok(users);
         }
 
+        [Authorize(Roles = "Admin")]
         [HttpGet("{id}")]
         public async Task<IActionResult> GetUserById(int id)
         {
@@ -51,26 +55,53 @@ namespace ChineseRaffleApi.Controllers
                 return BadRequest(ex.Message);
             }
         }
-
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateUser(int id, [FromBody] UpdateUserDto user)
         {
             try
             {
+                int userIdFromToken = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value
+                                                ?? throw new Exception("User ID not found in token"));
+
+                if (id != userIdFromToken)
+                    return Forbid("You are not allowed to update another user's data.");
+
                 await _userService.UpdateUserAsync(id, user);
                 return NoContent();
+            }
+            catch (FormatException ex)
+            {
+                return BadRequest($"Invalid user ID format: {ex.Message}");
             }
             catch (Exception ex)
             {
                 return BadRequest(ex.Message);
-            }         
+            }
         }
-
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteUser(int id)
         {
-            await _userService.DeleteUserAsync(id);
-            return NoContent();
+            try
+            {
+                int userIdFromToken = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value
+                                                ?? throw new Exception("User ID not found in token"));
+
+                if (id != userIdFromToken)
+                    return Forbid("You are not allowed to delete another user.");
+
+                await _userService.DeleteUserAsync(id);
+                return NoContent();
+            }
+            catch (FormatException ex)
+            {
+                return BadRequest($"Invalid user ID format: {ex.Message}");
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
     }
 }
